Added Winning project
Built the items and project logic, were some tiles have items that can be obtained after purchase, or after lading on the tile. If those items are required for the project, they will be added there. After all the items for the project are collected, the player wins. UI is required for this functionality, in addition to what happens after the player wins.